Use track info mutations
The problem with observing the forward button is that if user
unmutes during an ad then the next ad won't get muted because
forward button is already disabled and no mutations take place.
However track info usually changes after each ad so this should
work better but this is still not perfect.